Audio HAL: Add API to attach an effect to a device
Add a function to audio HAL interface allowing the attachement of an
audio effect to an audio device.
This is used when an audio effect is implemented below the HAL (e.g by an audio DSP)
and is attached/enabled when a particular sink(e.g speaker) or source(e.g mic) device
is selected.
This function must be supported from legacy HAL version 3.1.

Bug: 136294538
Test: make
Change-Id: Id63c83403fef580579c52d92758a552d67c3faec